algodjango/requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
Django==3.1.1
djangorestframework==3.11.1
Fiona==1.8.17
geopandas==0.8.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
munch==2.5.0
networkx==2.5
numpy==1.19.2
osmnx==0.16.0
pandas==1.1.2
Pillow==7.2.0
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
Rtree==0.9.4
Shapely==1.7.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10